requirements.txt
fastapi
google.generativeai
pydantic